pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "founder-endurance-env"
version = "1.0.0"
description = "An RL environment simulating a startup founder managing burnout, cash runway, and product velocity."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic",
    "numpy",
    "openai"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["models", "inference", "client"]

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]

[tool.setuptools.package-data]
"server" = ["*.py"]